Add "half" to geometric functions
We seem to have unintentionally omitted `half` support from the
Geometric Functions.  These functions derive from OpenCL, which
does have "half" variations.  We added `half` to all the other
builtin functions, so we just seem to have missed the Geometric
ones.

Note that the `fast_distance`, `fast_length`, and `fast_normalize`
functions do NOT have `half` variants (or `double` variants).  This
is intentional, and it matches OpenCL.  The rationale is that these
three functions all rely on the half-precision functions, which are
only defined for `float`.

Closes internal issue 660.