Fix definition of "nextafter" for double and half
The `nextafter` function returns the next floating-point value that is
representable in the given type.  The previous description was wrong
because it specifically stated that it returns the next
"single-precision" value.

I think we adopted the previous wording from OpenCL, and the OpenCL
committee recently fixed this same wording in their spec.

Closes internal issue 661.