Align description of logical `vec` operator overloads
The code snippet for `vec` interface does not list any restrictions on
`operator&&` and `operator||` and they are not listed in description of
overloads `(vec, vec)` and `(vec, scalar)`.

However, they were listed in description of overloads for `(scalar,
vec)` and there is no reasons for them to be there.

This commit aligns description of all overloads of logical `vec`
operators by removing data types restriction for
`vec::operator&&(scalar, vec)` and `vec::operator||(scalar, vec)`.