Use hyphens in reference names
Some Asciidoc references to the glossary terms were using the name of
the glossary term instead of the link name.  For glossary terms that
have multiple words, this results in a reference that has a space in
the name.  This space is causing problems in #481.

All these glossary terms have an anchor that is the same as the
glossary term, but with a hyphen instead of the space that separates
the words.  Change the references to use this anchor name instead of
using the glossary term.