Update copyright notice to revision 10
Update the copyright notice to the latest Khronos template, which is
revision 10.  Also update the copyright date to the current year.

This commit also fixes an unresolved issue with the previous notice by
adding a reference to the section that describes how to identify the
non-normative parts of the spec.