Clarify group_barrier synchronization behavior
The previous description of group_barrier required implementations to
execute certain fences, which should be an implementation detail.

This commit adds a "Synchronization" section to group_barrier and describes
behavior in terms of a happens-before relationship. Implementations may
choose to implement the barrier using any synchronization operations which
satisfy this requirement.