Builtin function precision is backend specific
Resolve conflicting wording about the precision requirements of the
builtin math functions.  The introductory paragraph of section 4.17
"SYCL built-in functions for SYCL host and device" says that the
precision is defined by the backend.  However, the descriptions in
section 4.17.4 "Math functions" and 4.17.6 "Half precision math
functions" specified a required precision for all backends.  Resolve
this by removing the the wording in 4.17.4 and 4.17.6.  (See issue #507
for more on this.)

Also add a new section to the OpenCL backend specification stating that
the precision of these builtin math functions for the OpenCL backend
matches the precision of the corresponding functions in the OpenCL
specification.

There are a few other related changes in this PR:

* Clarify that the native precision and half precision functions are
  available in both host and device code (see issue #509).

* Remove a statement that the geometric functions use round to nearest
  even rounding mode.  This also seems backend specific, and it doesn't
  even make sense for the OpenCL backend.  See OpenCL issue
  KhronosGroup/OpenCL-Docs#1024 for more details on this.

* Editorial changes to remove gratuitous references to OpenCL in the
  introductory paragraphs.

* Editorial changes to make the introductions of sections 4.17.4 -
  4.17.10 read consistently.

Closes #507
Closes #509