Clarify __swizzled_vec__ class
Prior to this commit, the `__swizzled_vec__` class was not precisely
specified.  We said that it had the same interface as `vec` except
for a small set of differences that we enumerated.  However, it became
clear after discussing with the original authors that this wasn't a
complete list of differences.

This commit adds a complete list of the members and friend functions
for the `__swizzled_vec__` type.  In many cases, the descriptions
simply defer to the `vec` specification, but there are also many cases
where the semantics are special for `__swizzled_vec__`.

The updated specification also describes `__swizzled_vec__` as a view
over the original `vec` object, which I think is what we originally
intended.  I intentionally removed wording like "may not be bound to a
l-value or escape the expression it was constructed in."  I think the
true restriction is that the lifetime of the `__swizzled_vec__` must
not outlive the lifetime of the underlying `vec`, which is the case
for any view.

I also intentionally removed the statement:

> Both kinds of swizzle member functions must not perform the swizzle
> operation themselves, instead the swizzle operation must be performed
> by the returned instance of `__swizzled_vec__` when used within an
> expression, meaning if the returned `__swizzled_vec__` is never used
> in an expression no swizzle operation is performed.

This wording made is sound like the point of `__swizzled_vec__` was to
delay the swizzle operation as a sort of optimization.  However, our
original intent was only to specify `__swizzled_vec__` as a view.

This commit also uses our updated style to specify the
`__swizzled_vec__` type, using a format that gives more horizontal
space for the synopses and descriptions.

Closes #504