Clarify requirements for some group functions
The group_broadcast and select_from_group functions are unlike the others, in
that they conditionally exchange values between work-items in the group. Some
developers may assume that it is safe for input arguments to be uninitialized
as long as conditions are such that those same values are not exchanged with
another work-item.

However, using an uninitialized value in this way results in undefined behavior
(according to C++), which may result in compilers performing optimizations that
are incompatible with the semantics of group functions.

This commit clarifies that the arguments passed to group_broadcast and
select_from_group must be initialized on all work-items. Although this
clarification is arguably unnecessary (since it is already covered by C++
rules), this is a common enough mistake that stating the precondition
explicitly may help developers who are new to SYCL.