Use fixed-width type in nan()
We noticed during implementation of the `nan` builtin function that the
specification did not ensure the `nancode` parameter had enough bits to
cover the entire range of the FP significand.  In particular,
`unsigned long` is not guaranteed to be wide enough for the 52 bits of
significand that are present in `double`.

It seems better to use the fixed-width integer types in this API, which
is what this commit proposes.