First step of updating the copyright dates
Add a script "update-copyright.sh" that automatically updates the
copyright dates, which was mostly copied from Ronan's PR #339.
Manually change the copyright line for a few files that don't follow
our regular pattern.

The file "copyright-ccby.txt" seems to be copied from the Vulkan repo.
I think we don't even use it, but the copyright notice listed
"Khronos Group" instead of "The Khronos Goup".  I replaced this file
with the latest version from the Vulkan repo, which has the corrected
name.