Clarify access::target::host_task
An accessor with access::target::host_task is accessible on
the host from within the host task lambda. It is necessary
to specify that no get_native_mem method is needed and also
that target::host_task means not only 'accessible within host_task'
but 'accessible within host_task on the host'. This is necessary
as host_task defines a scope, but it doesn't define where the
memory is accessible from.

It also adds reference to deprecated targets for accessors:
target::local and target::constant_buffer for accessors.