Change table style
We used to enclose all the member function descriptions in a table, so
that the description and synopsis of each function was visually
enclosed in a single table cell.  This caused a problem with the PDF
render, though, because some table rows were too big to fit on a single
page, which causes an error when rendering the PDF.  Fix this by
eliminating the table.  Instead, put a horizontal line between the
descriptions of the functions to visually separate them.