Fix typo with "accessor::get_pointer"
It appears that we introduced a typo with #432 with the return type of
`accessor::get_pointer`.  In the synopsis, we specified it as:

```
global_ptr<DataT> get_pointer() const noexcept
```

But in the description, it seems that we made a typo and said:

```
global_ptr<access::decorated::legacy> get_pointer() const noexcept
```

I believe these are both wrong, and it should be:

```
global_ptr<value_type> get_pointer() const noexcept
```

Using `value_type` is better than `DataT` because `value_type` is const
when the accessor is read-only.