Allow raw pointers in addition to multi_ptr
Several functions accepted `multi_ptr` as a parameter type but did not
accept raw pointer.  This was confusing to users, especially in cases
where the `multi_ptr` was allowed to be in `generic_space`.  Since raw
pointers behave like generic space multi-pointers, there's no reason
to disallow raw pointers for these functions.

Discussing this within our team, we felt this was an oversight, and not
something we intentionally left out.  Therefore, we think it is
appropriate as a bug fix to SYCL 2020.

This commit expands the parameter types allowed by these functions to
include raw pointers.

I did *not* change `async_work_group_copy` because this function takes
`multi_ptr` parameters in a specific address space (which did not
include `generic_space`).  Therefore, allowing raw pointers for this
function seemed inappropriate.