Recombine description of smoothstep
This is also a purely editorial change.  The specification of
`smoothstep` used to be split into two table rows only because the
definition was too long to fit on a single page.  (The PDF rendering
tools give an error when a table row is taller than a page.)  Now that
we no longer enclose the definition in a table, both overloads of
`smoothstep` can be described together in a single synopsis.