Expandable TOC for HTML output
Add an Asciidoctor extension and CSS styling to implement an expandable
table of contents for the HTML render.  This allows the user to click
an icon in the TOC to expand or collapse each heading level.

There is also an icon on the title line of the TOC which the user can
click to expand or collapse all heading levels.

The initial state is that all TOC levels are collapsed, but this is
tunable via the `toclevels-expanded` Asciidoc attribute.