Add automatic API cross references
Add an extension which creates automatic links from `[api]#NAME#` to
`[apidef]#NAME#`.  The `[api]` role is styled the same as `[code]`,
which means that we can replace some instances of `[code]` with `[api]`
when that role is used to refer to a SYCL identifier.  By using `[api]`
the identifier will also be a link to the synopsis that defines the
identifier.

Also change some instances of `[code]` to `[api]` in the math builtin
sections to make use of this new ability.