Clarify wording about local_access "value_type"
Because there is no access mode for `local_accessor`, the `value_type`
is always the same as `DataT`.  (The only way to get a read-only
`local_accessor` is to use a `DataT` that is `const`.)

Remove a comment in the synopsis that made it appear that `value_type`
might be different from `DataT`.

Also change the return type of `local_accessor::get_pointer` from
`DataT` to `value_type`.  Although these types are the same, it seems
clearer to be consistent with `accessor::get_pointer`, where the return
type is `value_type`.