Change the base C++ version
Change wording to allow the core C++ language to be any version C++17 or
later.

Move sections about normative references and non-normative wording to
the introduction, so they are earlier in the document.

My longer term goal is to remove "chapter/references.adoc" and add any
references we require to the "Normative references" section.  Some more
work needs to be done for this to happen:

* Decide if we need to list the C++ defect reports as normative
  references.  Currently, the "references" appendix lists only DR2325.
  Is there a reason to list this one and not others?

* Decide if OpenCL should be a normative reference.  If so, which
  version(s)?