Add placeholder stating there are no KHRs yet
It seems confusing to have an appendix with no KHR extensions.  I think
we will have at least one by the time this spec is published, but just
to be safe add a sentence stating that there are no KHR extensions yet
in the appendix.  We'll remove this sentence when we have our first KHR.