Document swizzle destructors
Add a section documenting the destructors for the
`__writeable_swizzle__` and `__const_swizzle__` classes.

I debated whether we should require them to be trivial.  I think they
probably will be trivial in a typical implementation, but I wasn't
certain enough to put this in the spec.  For now, I just said that they
have no visible effect.  This allows an implementation to contain some
internal state that is destroyed by the destructor, but I suspect most
implementations won't require this.