Device selector might be called more than once
The description of device selectors in the "Device selection" section
does not say that a selector is called exactly once for each root
device.  Change the wording of the `platform` selector constructor to be
consistent with this wording, leaving the number of times it is called
ambiguous.

We might want to reconsider this, though.  Why wouldn't we want to
require the selector to be called exactly once?