Fix Asciidoc error in vec table
It appears that this error was introduced a while ago in #480, but it
was only flagged recently as an error in the build.  The docker image
we use to build the spec was updated earlier today, and the new image
flags this as an error.  It appears that older docker images did not
flag an error, but the build output was corrupt.  Luckily, we did not
publish any of the corrupted builds.