Correct commit id when building from CI
The revision line at the top of the spec is supposed to include the
git commit ID that the spec was built from.  This works fine when
building manually, but there has been a longstanding bug preventing
this form working when the spec is built in GitHub CI.

I think the problem is that the GitHub action doesn't check out any of
the git history, so the "git describe" and "git log" commands don't
work.  Rather than use git commands to get the commit ID, we can just
get the ID from the GitHib $GITHUB_SHA variable.

I also simplified the information in the revision line, so it just
contains the commit ID and not the branch information.  I think this is
the only critical information because we can always map the ID back to a
branch later if we want.