Clarify backend and introduce "backend::none"
Clarify exactly what we mean by "backend".  I think it has always been
our intention that the `backend` enumeration and the `get_backend`
member function tell an application what sort of interoperation is
supported by a SYCL object.  For example, if a SYCL object returns
`backend::opencl` from its `get_backend` function, that SYCL object
supports interoperation as defined by the OpenCL backend interoperation
specification.

This commit clarifies the spec to say that explicitly.

It also became clear that we need a `backend::none` enumerator to
indicate that a SYCL object does not support any backend interoperation.
For example, this would be useful if a vendor implements SYCL directly
on hardware, without using a documented low-level offload API.  Many of
the changes in this commit merely remove sentences that imply that every
SYCL object necessarily has some backend.

This commit also makes a minor change to the default `event`
constructor.  Previously, we required the backend for such an event to
be the same as the backend for the default device.  After some
implementation experience, we decided that this is not practical.
Requiring interoperation with a particular backend constrains the
implementation too much.  It also seems arbitrary to require the
default-constructed event to have a particular backend.  To address
these concerns, this commit loosens the requirement, allowing an
implementation to choose which backend (if any) the default-constructed
event has.

Closes #564