Go back to referencing by SHA
It appears that it is legal to reference the docker image using both
its tag and its SHA, but the tag is totally ignored in that case.
Since there is no error checking that the tag matches the SHA, it seems
better to just use the SHA and rely on a comment to document which tag
this corresponds to.