Fix local docker build of PDF spec
Something changed in the latest docker image that causes asciidoctor-pdf
to fail when run locally (but not when run in CI).  The problem is that
the user is different when running locally, and this user does not have
write access to the HOME directory.  It appears that asciidoctor-pdf
attempts to write temporary files into HOME, so this failed with an
access permission error.

Fix this by setting HOME=/tmp when doing a local docker build.