Relax backend for default-constructed event
Relax the requirements on the backend for a default-constructed `event`
object.  Previously, we required the backend for such an event to be the
same as the backend for the default device.  After some implementation
experience, we decided that this is not practical.  Requiring
interoperation with a particular backend constrains the implementation
too much.  It also seems arbitrary to require the default-constructed
event to have a particular backend.  To address these concerns, we now
say that the backend that is associated with a default-constructed
event is implementation defined.