Disallow repeated swizzle element in assignment
We previously decided that assignment to a `__writeable_swizzle__`
should have a constraint that the lhs does not have any repeated
swizzle elements.  Thus, something like this is disallowed:

```
vec<int, 2> x;
myvec.swizzle<1,1>() = x;
```

We did this because the result would depend on the order in which the
assignments occurred, and we think that some compilers will not handle
this case well.

However, we neglected to add the same constraint to other
assignment-like swizzle operations.  This commit adds a similar
constraint, which disallows all of the following cases also, where
the `__writeable_swizzle__` has repeated elements:

```
myvec.swizzle<1,1>().load(off, p);
myvec.swizzle<1,1>() += x;
myvec.swizzle<1,1>()++;
++(myvec.swizzle<1,1>());
```