Undeprecate legacy multi_ptr
The legacy decoration for multi_ptr has been made the default template
argument of the decoration in multi_ptr to bridge the gap between
SYCL 1.2.1 and SYCL 2020 multi_ptr interfaces. However, the legacy
decoration and the specialization of multi_ptr with the decoration is
deprecated with SYCL 2020. This is confusing to users as they may easily
run into deprecation warnings simply due to not specifying the
decoration. This commit proposes that we remove the deprecation of
legacy in SYCL 2020.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>