Improve wording about returned event
Improve the wording of the event that is returned by `queue::submit` and
friends.

Since the new wording says that the event represents the *command* that
is submitted to the queue, it seemed awkward for the _Effects_ paragraph
to say that a *command group function* is submitted to the queue.  I
replaced this with wording saying that the command group function is
immediately called, which is consistent with what we say in the
introductory paragraphs of section 4.6.5 "Queue class".  The new wording
also notes that the command group function may submit no more than one
command to the queue, which is consistent with the existing wording in
section 4.9.3 "Command group scope".