Clarify and deprecate "execution_capabilities"
We had a customer bug report about the behavior of
`info::device::execution_capabilities` for non-OpenCL devices.  It turns
out that the spec isn't clear about this.  This PR clarifies that an
exception should be thrown, which is consistent with our behavior for
`info::device::profile` and `info::device::preferred_interop_user_sync`,
which are also inherited from OpenCL.  Another option would be to return
an empty vector, so we could debate that if people think that's better.

I also think we should deprecate this query, which is consistent with
the other queries that are specific to the OpenCL backend.  This PR
deprecates the query in SYCL-NEXT, though I would also be open to
retroactively deprecating it in SYCL 2020 with the rationale that we
deprecated the other OpenCL-specific queries and simply forgot this one.