Fix contradiction with context constructors
The context constructor overloads in the `context` synopsis did not
match the description in the text of the specification.  We decided that
the overloads in the synopsis were what we intended, so change the text
to match.

We also discovered that two constructor overloads taking a `platform`
argument were present in SYCL 1.2.1 and were inadvertently dropped from
the SYCL 2020 specification.  Add them back.

Clarify the behavior of the constructors when the set of devices is
empty.  We decided that this should throw an exception, which is
consistent with OpenCL's `clCreateContext`.

Closes #470
Closes #474