Fix error behavior for initialize_to_identity
Previously, the specification said that a compiler must issue a diagnostic if
the initialize_to_identity property was used without a known identity value.
This requirement cannot be satisfied, because the presence of the
initialize_to_identity value in a property_list cannot be proven until
run-time.

This change replaces the compile-time diagnostic with a run-time exception.