Fix vec constructors
This is change 1 of 9 that fix problems with the specification of the
`vec` class.  An implementation that follows the existing specification
would not accept common code patterns and would not pass the CTS.  None
of the existing implementations actually follow the existing
specification.

This change updates the constructors:

* Allow the variadic constructor parameters to be types implicitly
  convertible to `DataT`.
* Allow the variadic constructor parameters to be the swizzle types.
* Allow implicit conversions from scalar type `T` to 1-element `vec` where
  `T` is implicitly convertible to `DataT`.

These changes correspond to slides 3 - 7 of the presentation that was
discussed in the WG meetings