Clarify "vec" ADL behavior
This is change 4 of 9 that fix problems with the specification of the
`vec` class.  An implementation that follows the existing specification
would not accept common code patterns and would not pass the CTS.  None
of the existing implementations actually follow the existing
specification.

This change clarifies the ADL behavior of the `vec` class.  Without
this clarification, implementations could make additional operators
available on the ADL search set, which could lead to ambiguity errors
when compiling applications.  For example, we know that adding `vec`
to the ADL search set for the swizzled vector operators leads to
ambiguity errors.

These changes correspond to slide 14 of the presentation that was
discussed in the WG meetings.