Clarify "vec" DataT
This is change 6 of 9 that fix problems with the specification of the
`vec` class.  An implementation that follows the existing specification
would not accept common code patterns and would not pass the CTS.  None
of the existing implementations actually follow the existing
specification.

This change clarifies the set of legal data types for a `vec` element
and also clarifies what `sycl::byte` is.

These changes correspond to slides 23 - 25 and 28 of the presentation
that was discussed in the WG meetings.